tests/wpt/tests/css/css-syntax/charset/xml-stylesheet-page-windows-1251-charset-attribute-windows-1250.xhtml

<?xml version="1.0" encoding="windows-1251"?>
<?xml-stylesheet href="support/no-decl.css" charset="windows-1250"?>
<html xmlns="http://www.w3.org/1999/xhtml"><head>
<title>CSS charset: page windows-1251, charset attribute bogus</title>
<link rel="help" href="https://drafts.csswg.org/css-syntax-3/#determine-the-fallback-encoding"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head><body>
<div id="log"></div>
<div id="foo"></div>
<script>
var elm = document.getElementById('foo');
elm.id = '\u010C';
var t = async_test();
onload = t.step_func(function(){
  assert_equals(getComputedStyle(elm, '').visibility, 'hidden');
  this.done();
});
</script>
</body></html>
